+++ main/src/main/java/net/oneandone/stool/stage/Stage.java
@@ -529,7 +529,7 @@ public abstract class Stage {
     private void checkMemory() throws IOException {
         int requested;
 
-        requested = configuration.tomcatHeap;
+        requested = configuration.memory;
         int unreserved = session.memUnreserved();
         if (requested > unreserved) {
             throw new ArgumentException("Cannot reserve memory:\n"
@@ -619,8 +619,8 @@ public abstract class Stage {
     //--
 
     public void tuneConfiguration() throws IOException {
-        if (configuration.tomcatHeap == 0 || configuration.tomcatHeap == 350) {
-            configuration.tomcatHeap = Math.min(4096, 150 + size() * session.configuration.baseHeap);
+        if (configuration.memory == 0 || configuration.memory == 400) {
+            configuration.memory = Math.min(4096, 200 + size() * session.configuration.baseHeap);
         }
         if (configuration.build.isEmpty() || configuration.build.equals("false")) {
             configuration.build = getDefaultBuildCommand();